+++ tests/unit/test_charm_state.py
@@ -9,7 +9,7 @@ from unittest.mock import MagicMock
 import pytest
 import yaml
 from charms.data_platform_libs.v0.data_interfaces import DatabaseRequires
-from github_runner_manager.types_.github import GitHubOrg, GitHubRepo
+from github_runner_manager.configuration.github import GitHubOrg, GitHubRepo
 from pydantic import BaseModel
 from pydantic.error_wrappers import ValidationError
 from pydantic.networks import IPv4Address
@@ -593,7 +593,7 @@ def test_proxy_config_from_charm(
     mock_charm.config[USE_APROXY_CONFIG_NAME] = False
     monkeypatch.setattr(charm_state, "get_env_var", MagicMock(side_effect=[http, https, no_proxy]))
 
-    result = ProxyConfig.from_charm(mock_charm)
+    result = charm_state._build_proxy_config_from_charm(mock_charm)
 
     assert result.no_proxy is None
 
@@ -803,12 +803,8 @@ class MockModel(BaseModel):
 @pytest.mark.parametrize(
     "module, target, exc",
     [
-        (
-            ProxyConfig,
-            "from_charm",
-            ValidationError([], MockModel),
-        ),
-        (ProxyConfig, "from_charm", ValueError),
+        (charm_state, "_build_proxy_config_from_charm", ValidationError([], MockModel)),
+        (charm_state, "_build_proxy_config_from_charm", ValueError),
         (CharmConfig, "from_charm", ValidationError([], MockModel)),
         (CharmConfig, "from_charm", ValueError),
         (charm_state, "_get_supported_arch", UnsupportedArchitectureError(arch="testarch")),
@@ -825,7 +821,7 @@ def test_charm_state_from_charm_invalid_cases(
     """
     mock_charm = MockGithubRunnerCharmFactory()
     mock_database = MagicMock(spec=DatabaseRequires)
-    monkeypatch.setattr(ProxyConfig, "from_charm", MagicMock())
+    monkeypatch.setattr("charm_state._build_proxy_config_from_charm", MagicMock())
     mock_charm_config = MagicMock()
     mock_charm_config.openstack_clouds_yaml = None
     mock_charm_config_from_charm = MagicMock()
@@ -848,7 +844,7 @@ def test_charm_state_from_charm(monkeypatch: pytest.MonkeyPatch):
     """
     mock_charm = MockGithubRunnerCharmFactory()
     mock_database = MagicMock(spec=DatabaseRequires)
-    monkeypatch.setattr(ProxyConfig, "from_charm", MagicMock())
+    monkeypatch.setattr("charm_state._build_proxy_config_from_charm", MagicMock())
     monkeypatch.setattr(CharmConfig, "from_charm", MagicMock())
     monkeypatch.setattr(OpenstackRunnerConfig, "from_charm", MagicMock())
     monkeypatch.setattr(charm_state, "_get_supported_arch", MagicMock())
